Add support for building only editor or only templates

+++ build-linux/build.sh
@@ -20,29 +20,37 @@ if [ "${CLASSICAL}" == "1" ]; then
 
   export PATH="${GODOT_SDK_LINUX_X86_64}/bin:${BASE_PATH}"
 
-  $SCONS platform=x11 $OPTIONS tools=yes target=release_debug
-  mkdir -p /root/out/x64/tools
-  cp -rvp bin/* /root/out/x64/tools
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86_64}" == "1" ]; then
+    $SCONS platform=x11 $OPTIONS tools=yes target=release_debug
+    mkdir -p /root/out/x64/tools
+    cp -rvp bin/* /root/out/x64/tools
+    rm -rf bin
+  fi
 
-  $SCONS platform=x11 $OPTIONS tools=no target=release_debug
-  $SCONS platform=x11 $OPTIONS tools=no target=release
-  mkdir -p /root/out/x64/templates
-  cp -rvp bin/* /root/out/x64/templates
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=x11 $OPTIONS tools=no target=release_debug
+    $SCONS platform=x11 $OPTIONS tools=no target=release
+    mkdir -p /root/out/x64/templates
+    cp -rvp bin/* /root/out/x64/templates
+    rm -rf bin
+  fi
 
   export PATH="${GODOT_SDK_LINUX_X86}/bin:${BASE_PATH}"
 
-  $SCONS platform=x11 $OPTIONS tools=yes target=release_debug bits=32
-  mkdir -p /root/out/x86/tools
-  cp -rvp bin/* /root/out/x86/tools
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86}" == "1" ]; then
+    $SCONS platform=x11 $OPTIONS tools=yes target=release_debug bits=32
+    mkdir -p /root/out/x86/tools
+    cp -rvp bin/* /root/out/x86/tools
+    rm -rf bin
+  fi
 
-  $SCONS platform=x11 $OPTIONS tools=no target=release_debug bits=32
-  $SCONS platform=x11 $OPTIONS tools=no target=release bits=32
-  mkdir -p /root/out/x86/templates
-  cp -rvp bin/* /root/out/x86/templates
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=x11 $OPTIONS tools=no target=release_debug bits=32
+    $SCONS platform=x11 $OPTIONS tools=no target=release bits=32
+    mkdir -p /root/out/x86/templates
+    cp -rvp bin/* /root/out/x86/templates
+    rm -rf bin
+  fi
 fi
 
 # Mono
@@ -57,30 +65,38 @@ if [ "${MONO}" == "1" ]; then
   export PATH="${GODOT_SDK_LINUX_X86_64}/bin:${BASE_PATH}"
   export OPTIONS_MONO_PREFIX="${OPTIONS} ${OPTIONS_MONO} mono_prefix=${MONO_PREFIX_X86_64}"
 
-  $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes
-  mkdir -p /root/out/x64/tools-mono
-  cp -rvp bin/* /root/out/x64/tools-mono
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86_64}" == 1 ]; then
+    $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes
+    mkdir -p /root/out/x64/tools-mono
+    cp -rvp bin/* /root/out/x64/tools-mono
+    rm -rf bin
+  fi
 
-  $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release_debug
-  $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release
-  mkdir -p /root/out/x64/templates-mono
-  cp -rvp bin/* /root/out/x64/templates-mono
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == 1 ]; then
+    $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release_debug
+    $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release
+    mkdir -p /root/out/x64/templates-mono
+    cp -rvp bin/* /root/out/x64/templates-mono
+    rm -rf bin
+  fi
 
   export PATH="${GODOT_SDK_LINUX_X86}/bin:${BASE_PATH}"
   export OPTIONS_MONO_PREFIX="${OPTIONS} ${OPTIONS_MONO} mono_prefix=${MONO_PREFIX_X86}"
 
-  $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes bits=32
-  mkdir -p /root/out/x86/tools-mono
-  cp -rvp bin/* /root/out/x86/tools-mono
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86}" == 1 ]; then
+    $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes bits=32
+    mkdir -p /root/out/x86/tools-mono
+    cp -rvp bin/* /root/out/x86/tools-mono
+    rm -rf bin
+  fi
 
-  $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release_debug bits=32
-  $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release bits=32
-  mkdir -p /root/out/x86/templates-mono
-  cp -rvp bin/* /root/out/x86/templates-mono
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == 1 ]; then
+    $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release_debug bits=32
+    $SCONS platform=x11 $OPTIONS_MONO_PREFIX tools=no target=release bits=32
+    mkdir -p /root/out/x86/templates-mono
+    cp -rvp bin/* /root/out/x86/templates-mono
+    rm -rf bin
+  fi
 fi
 
 echo "Linux build successful"
