+++ build-windows/build.sh
@@ -18,27 +18,35 @@ source /root/common/prep.sh
 if [ "${CLASSICAL}" == "1" ]; then
   echo "Starting classical build for Windows..."
 
-  $SCONS platform=windows bits=64 $OPTIONS tools=yes target=release_debug
-  mkdir -p /root/out/x64/tools
-  cp -rvp bin/* /root/out/x64/tools
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86_64}" == "1" ]; then
+    $SCONS platform=windows bits=64 $OPTIONS tools=yes target=release_debug
+    mkdir -p /root/out/x64/tools
+    cp -rvp bin/* /root/out/x64/tools
+    rm -rf bin
+  fi
 
-  $SCONS platform=windows bits=64 $OPTIONS tools=no target=release_debug
-  $SCONS platform=windows bits=64 $OPTIONS tools=no target=release
-  mkdir -p /root/out/x64/templates
-  cp -rvp bin/* /root/out/x64/templates
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=windows bits=64 $OPTIONS tools=no target=release_debug
+    $SCONS platform=windows bits=64 $OPTIONS tools=no target=release
+    mkdir -p /root/out/x64/templates
+    cp -rvp bin/* /root/out/x64/templates
+    rm -rf bin
+  fi
 
-  $SCONS platform=windows bits=32 $OPTIONS tools=yes target=release_debug
-  mkdir -p /root/out/x86/tools
-  cp -rvp bin/* /root/out/x86/tools
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86}" == "1" ]; then
+    $SCONS platform=windows bits=32 $OPTIONS tools=yes target=release_debug
+    mkdir -p /root/out/x86/tools
+    cp -rvp bin/* /root/out/x86/tools
+    rm -rf bin
+  fi
 
-  $SCONS platform=windows bits=32 $OPTIONS tools=no target=release_debug
-  $SCONS platform=windows bits=32 $OPTIONS tools=no target=release
-  mkdir -p /root/out/x86/templates
-  cp -rvp bin/* /root/out/x86/templates
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=windows bits=32 $OPTIONS tools=no target=release_debug
+    $SCONS platform=windows bits=32 $OPTIONS tools=no target=release
+    mkdir -p /root/out/x86/templates
+    cp -rvp bin/* /root/out/x86/templates
+    rm -rf bin
+  fi
 fi
 
 # Mono
@@ -52,29 +60,37 @@ if [ "${MONO}" == "1" ]; then
 
   export OPTIONS_MONO_PREFIX="${OPTIONS} ${OPTIONS_MONO} mono_prefix=${MONO_PREFIX_X86_64}"
 
-  $SCONS platform=windows bits=64 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes
-  mkdir -p /root/out/x64/tools-mono
-  cp -rvp bin/* /root/out/x64/tools-mono
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86_64}" == "1" ]; then
+    $SCONS platform=windows bits=64 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes
+    mkdir -p /root/out/x64/tools-mono
+    cp -rvp bin/* /root/out/x64/tools-mono
+    rm -rf bin
+  fi
 
-  $SCONS platform=windows bits=64 $OPTIONS_MONO_PREFIX tools=no target=release_debug
-  $SCONS platform=windows bits=64 $OPTIONS_MONO_PREFIX tools=no target=release
-  mkdir -p /root/out/x64/templates-mono
-  cp -rvp bin/* /root/out/x64/templates-mono
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=windows bits=64 $OPTIONS_MONO_PREFIX tools=no target=release_debug
+    $SCONS platform=windows bits=64 $OPTIONS_MONO_PREFIX tools=no target=release
+    mkdir -p /root/out/x64/templates-mono
+    cp -rvp bin/* /root/out/x64/templates-mono
+    rm -rf bin
+  fi
 
   export OPTIONS_MONO_PREFIX="${OPTIONS} ${OPTIONS_MONO} mono_prefix=${MONO_PREFIX_X86}"
 
-  $SCONS platform=windows bits=32 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes
-  mkdir -p /root/out/x86/tools-mono
-  cp -rvp bin/* /root/out/x86/tools-mono
-  rm -rf bin
+  if [ "${BUILD_EDITOR_x86}" == "1" ]; then
+    $SCONS platform=windows bits=32 $OPTIONS_MONO_PREFIX tools=yes target=release_debug copy_mono_root=yes
+    mkdir -p /root/out/x86/tools-mono
+    cp -rvp bin/* /root/out/x86/tools-mono
+    rm -rf bin
+  fi
 
-  $SCONS platform=windows bits=32 $OPTIONS_MONO_PREFIX tools=no target=release_debug
-  $SCONS platform=windows bits=32 $OPTIONS_MONO_PREFIX tools=no target=release
-  mkdir -p /root/out/x86/templates-mono
-  cp -rvp bin/* /root/out/x86/templates-mono
-  rm -rf bin
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=windows bits=32 $OPTIONS_MONO_PREFIX tools=no target=release_debug
+    $SCONS platform=windows bits=32 $OPTIONS_MONO_PREFIX tools=no target=release
+    mkdir -p /root/out/x86/templates-mono
+    cp -rvp bin/* /root/out/x86/templates-mono
+    rm -rf bin
+  fi
 fi
 
 echo "Windows build successful"
